# Slightly faster handling of odd byte numbers