I'm having a bit of trouble making the code as fast for byte numbers not a multiple of 8, which is odd, since e.g. 23 bytes should never be slower than 24. E.g. I currently have
```
Small key speed test -   22-byte keys -    26.00 cycles/hash
Small key speed test -   23-byte keys -    27.75 cycles/hash
Small key speed test -   24-byte keys -    22.00 cycles/hash
Small key speed test -   25-byte keys -    24.00 cycles/hash
Small key speed test -   26-byte keys -    25.00 cycles/hash
```
Anyway, this patch makes it slightly better on my system at least.