@rurban 2020-09-03
Unaligned rest access willl be slower on almost all modern platforms. At the beginning and at the end. This is the rest at the end. Looks right to me.